requirements/development.txt
-r base.txt

# Jakość kodu
black>=24.3.0
flake8>=7.0.0
isort>=5.13.2
mypy>=1.9.0
django-stubs>=4.2.7
pre-commit>=3.7.0

# Django narzędzia deweloperskie
django-debug-toolbar>=4.3.0
django-extensions>=3.2.3

# Testowanie
pytest-django>=4.8.0
factory-boy>=3.3.0
Faker>=24.3.0
coverage>=7.4.3

# Scraping RRUP (jednorazowo)
playwright>=1.44.0
beautifulsoup4>=4.12.3

# Narzędzia
ipython>=8.23.0
rich>=13.7.1
